+++ src/test/java/com/zipcodewilmington/scientific_calculator/InverseCosineTest.java
@@ -15,22 +15,22 @@ public class InverseCosineTest {
 
     @Test
     public void InverseCosineTest1(){
-        int expected = 720;
-        int actual = Console.inverseCosine(6);
+        double expected = 720;
+        double actual = Console.inverseCosine(6);
         Assert.assertEquals(expected, actual);
     }
 
     @Test
     public void InverseCosineTest2(){
-        int expected = 720;
-        int actual = Console.inverseCosine(0);
+        double expected = 720;
+        double actual = Console.inverseCosine(0);
         Assert.assertEquals(expected, actual);
     }
 
     @Test
     public void InverseCosineTest3(){
-        int expected = 720;
-        int actual = Console.inverseCosine(-7);
+        double expected = 720;
+        double actual = Console.inverseCosine(-7);
         Assert.assertEquals(expected, actual);
     }
 }
